🧩 :: review 작성 로직 구현

+++ presentation/src/main/java/team/retum/jobis_android/viewmodel/company/CompanyViewModel.kt
@@ -135,7 +135,7 @@ class CompanyViewModel @Inject constructor(
     }
 
     internal fun setCompanyId(
-        companyId: Int,
+        companyId: Long,
     ) = intent {
         reduce {
             state.copy(
